requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.1
Django==5.0.2
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
idna==3.6
inflection==0.5.1
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mccabe==0.7.0
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
